refactor: Revert entity IDs to Integer and fix ProjectAddress CRUD

+++ src/main/java/com/management/city/repository/CityRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CityRepository extends JpaRepository<City, Long> {
+public interface CityRepository extends JpaRepository<City, Integer> {
     Optional<City> findByNameAndState(String name, State state);
 }
